Web apis

+++ CatalogService/Catalog.Core/Services/CategoryService.cs
@@ -52,12 +52,24 @@ namespace Catalog.Core.Services
         /// <inheritdoc/>
         public async Task RemoveCategory(int id)
         {
+            var categoryFromDB = await this.repository.GetCategory(id);
+            if (categoryFromDB is null)
+            {
+                throw new NotFoundException($"Category with {id} id was not found.");
+            }
+
             await this.repository.RemoveCategory(id);
         }
 
         /// <inheritdoc/>
         public async Task UpdateCategory(Category category)
         {
+            var categoryFromDB = await this.repository.GetCategory(category.Id);
+            if (categoryFromDB is null)
+            {
+                throw new NotFoundException($"Category with {category.Id} id was not found.");
+            }
+
             if (category.ParentCategoryId is not null)
             {
                 var parentCategory = await this.repository.GetCategory(category.ParentCategoryId.Value);
